+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+* Add debug output.
+
 ## [0.1.0] - 2023-04-25
 
 ### Added